requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.1
gunicorn==20.1.0
numpy==1.21.2
opencv-python==4.5.3.56
dlib==19.22.1
cloudinary==1.26.0
python-dotenv==0.19.0